map-reduce/requirements.txt
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1596828493937/work
attrs @ file:///tmp/build/80754af9/attrs_1598374659300/work
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
cffi @ file:///tmp/build/80754af9/cffi_1598370769933/work
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1593446406207/work
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596207638929/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipympl==0.3.3
ipython @ file:///tmp/build/80754af9/ipython_1598883821407/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi @ file:///tmp/build/80754af9/jedi_1598371611696/work
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1597876319132/work
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mpi4py==3.0.3
nbconvert==5.6.1
nbformat==5.0.7
notebook @ file:///tmp/build/80754af9/notebook_1596838629813/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1596233721170/work
olefile==0.46
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1594307295532/work
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1598885458782/work
ptyprocess==0.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.6.1
Pympler @ file:///tmp/build/80754af9/pympler_1599767262062/work
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pyzmq==19.0.1
Send2Trash==1.5.0
sip==4.19.13
six==1.15.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
